feat(posts_controller.rb): add new and create actions to PostsController
The new action sets up a new Post object and assigns it to an instance variable. The create action creates a new Post object with the permitted parameters and assigns it to an instance variable. It then sets the user and clubhouse associations for the Post object. If the Post object is saved successfully, it updates the clubhouse posts list with a turbo stream and redirects to the clubhouse page. If the Post object is not saved successfully, it renders the new template with a notice.