feat(views): add conditional rendering of author information based on user membership
Add conditional rendering of author information based on user membership in the following views:
- app/views/clubhouses/show.erb
- app/views/partials/clubhouses/_card.html.erb
- app/views/partials/comments/_card.html.erb
- app/views/partials/posts/_card.html.erb
- app/views/posts/show.erb

If the user is not signed in, a message is displayed asking them to log in. If the user is signed in but not a member of the clubhouse, a message is displayed informing them that they must be a member of the clubhouse to see the author of the post or comment. If the user is signed in and a member of the clubhouse, the author information is displayed.